requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
rich>=13.0.0
markdown>=3.5.0
playwright>=1.40.0
weasyprint>=60.0
